:construction: create processor
