--- nonebot/dependencies/__init__.py
@@ -1,2 +1,0 @@
-from .models import Depends as Depends
-from .utils import get_dependent as get_dependent
